requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
python-multipart